Removed a bunch of cruft relating to the video language page
We had code to handle going to the "video history" page, without a
language.  I'm not sure why this made sense in the past, but it
doesn't make sense now and it wasn't used.  Removed a couple unused
functions/templates.  Changed the rest to simply redirect to the video
page, which is what we were doing anyways.