New system for tracking hit counts
Implemented a new way to track hit counts.  This one doesn't use
redit, it just inserts into a DB table.

The old way had a fairly complicated system for adding counts:
  - It would create a task to insert the count
  - The task would update a redis field
  - Every so often another task would migrate that field to the
    regular DB

The new system is:
  - Insert a row for the hit using INSERT DELAYED
  - Every day we count rows in the hit table for yesterday and
    aggregate them to the per-day table
  - At the start of the month we sum up counts in the per-day table
    for last month and move them to the per-month table.
  - We then delete hits older than 24 hours and per-day counts older
    than 1 month.

I think the new system should perform better than the old system.
Using INSERT DELAYED should be just as fast, if not faster than
scheduling a task.  The total number of rows in the 3 tables should be
reasonable (right now we have like 23 million rows in the per-day
table, which makes queries slow).  We avoid scheduling a task for each
hit, which was resulting in something like tens of thousands of tasks
per hour.