Added form to link a team with a kaltura account
Made a new app called externalsites.  It's going to be the new way to
handle syncing and related tasks.  It should replace the accountlinker
app.

Put the view code in externalsites for the teams accounts tab.  I
think it's better to add code to the new app rather than jamming more
stuff in the teams module.

Did some work to make it easy to add account forms for other sites as
we need them.