Started work on syncing subtitles to external sites
Listen to various signals and call a task when they fire.  The actual
tasks are still to be implemented.

Added some code to generalize account models.  Right now we just have
KalturaAccount, but as we add more we need a way to reference objects
from different model classes.  Implemented the simplest system to do
this, we use a type value and the object id.  Changed KalturaAccount
slightly so that we have an object id.