Fix embedder media compilation.
For some reason using the closure compiler wasn't working.  Switched
to simply concatinating the files like we do with the subtitle editor
and things seem to work again.

Fix for the include_bundle tag to work with the embedder JS.

I don't really understand this code at all, it seems like it could use
a big cleanup to me.