Bunch of fixes to the docker-dev stuff
Went away from trying to store imageids for all the images.  I realize
that those will change and there's not an easy way to implement what I
wanted.  Replaced it with a flag you can pall to startup.py

Detect when images are removed, but the cidfile still exists.  If that
happens we remove the cidfile and restart the image.

Made the amara-dev Dockerfile not depend on the main one.  I think
this is better since the main one installs a bunch of stuff that we
just replace when we mount the git repo in /opt/apps/unisubs.

Various fixes to the scripts/modules.