Implemented initial search code in the transcript viewer
Stopped using the amaratranscript popcorn plugin.  It was getting a
bit weird to my mind because some things were done in the plugin, some
things were done in the backbone view, sometimes the plugin would call
a method on the backbone view.  I think it's simpler to have the
plugin simply control the CSS class of the current line and the code
plugin work fine for that.

Refactor the way we initialize the backbone views/models to
consolidate things.  The new system is loadSubtitles ensures the
subtitles have been fetched for a given language, then calls
setCurrentLanguage().  This is nicer since all the setup code is in
one place.