Fix thumbnail uploads (#920)
The uploads were failing because we used a really old version of
sorl-thumbnail.  Replaced that with code that uses easy_thumbnails to
create the thumbnails.  Also replaced PIL with a recent version of
Pillow.

Removed the thumb_options parameter.  We weren't really using it.  Now
we always upscale and use crop=smart (which was the default before).