Don't make tab move the subtitle focus (#965)
The issue was that bound the tab key to 2 different actions:
play/pause and navigating to the next subtitle.  When the user hit it,
we were doing both.  In the case of 1 subtitle, we would just let that
subtitle lose focus.

Changed it so tab only toggles play/pause.  Also made the other
keybindings when editing subtitles (enter/escape), not propagate
upwards.  This way if we add a global keybinding for one of those, we
won't have both actions fire at once, which was a confusing part of
this issue.