Implemented saving task notes without approving/rejecting
Added code to display the task notes when the editor is first loaded.
Before we would only show them when we were coming from the old
editor.

Added code to track when notes change and enable the save button in
that case.

Updated SaveSessionController.SaveSession to handle saving notes if we
need to.  Refactored that function a bit to make full use of the
deferred/promise API.