Fixes for team video feeds (#837) (#836)
Refactored/cleaned up the code things.  Now the team code is in the
teams app and the videos code is in the videos app.  I also generally
tried to make things clearer.

Save the team with for video feeds added via the team page.  Create
team videos whenever that feed is updated (before we would just do it
on the first pass).