Cleaned up team videos filters CSS/HTML
Moved the buttons to before the filters, this makes things overflow
correctly.

Removed the margin-bottom from the filter-chunk and added it to
padding on the filters DIV.  This makes things wrap better when they
overflow.

Changed "Sorted by" to "Sort:" to match the other filters.

Made project/language each be in their own filter-chunk.  This allows
them to wrap separately if they need to.