Made it so searching doesn't clear out the other filters
Split up teams/_search.html template into the search form and the
javascript.  It seems cleaner for each template that's including that
to include their own forms, instead of forcing each to use the same
variable names.

Pass the language and project filters to the teams/videos-list.html
template.  That way we can continue to pass them on if the user
searches.

Went back to using "project" for the GET parameter, not
"project-filter".  I'm not sure what my intentions there were, but the
view code was still using project, so I figure we should keep that.