Fixed team ORM in the south database.
I'm not sure exactly what happened, but upgrading south to version 0.7
caused it to think that every BooleanField had changed.  I think
because it now stores blank=True, where before it ignored that.  In
any case, it doesn't matter at the mysql level, so I just added an
empty migration.