Reworked the collaborate section (#1046)
Removed it entirely if the user can't create/translate subtitles.
Reworded things slightly for teams that are moderated, but don't use
tasks.  Moved the code that checks permissions out of the template and
into the view.

Moved the template for team's that use tasks from _team_list.html to
the video-page.html template.  Changed it so that it only displays the
"Contribute subtitles: view tasks for this video".  Refactored the
template code to make it translatable. Removed the _team_list.html
template and the render_belongs_to_team_list tag.