More videos page fixes
Moved team info under the Contribute header Removed unneeded CSS after
this change.

Added whitespace above our description and metadata sections.  Made
the description DIV use the styles that it had before, for example
line-height: 150%.

Moved languages metrics box to be the 1st in the list.  Started using
table layouts for the metrics lists.  This is the easiest way to make
it render well regardless of the number of cells.

Added extra space between the various <p> elements in the Contribute
section.  Changed wording for the link to the team.