Fix subtitle download links (#929)
Changed it so instead of sending a Content-Disposition header, we
simply use the filename we want in the URL.  This works better with
non-ascii characters.

Moved the view from the widgets app to the subtitles app.

Cleaned up the widget javascript a bit.

Copied team member code from gh-1046.  I think this code is nicer than
the current stuff and I wanted to use it for the subtitles download
view.