Fix download urls for videos with a newline in the title (#1133)
Changed it so we use the video title rather than the language title
for download filenames.

Changed the way we handle download URLs for the widget.  Now we
calculate the filename in python land and pass it to the widget.  This
makes it so we don't have to write the same code twice to convert the
video title to a filename.