Refactored how we implement the subtitle-repeat directive.
Moved the code from the template to the directive itself.  Having it
the template was nice when it was just interpolations, but now that
it's getting more complex it's kind of crazy.  We're basically
re-inventing the angular template language and I'd rather not continue
to do that.

This should be functionally the same as before, but hopefully easier to
extend.

Updated/cleaned up the CSS and markup we use for each subtitle LI:
  - Use <button> instead of <a> tags
  - Just put the image inside of the button rather than the CSS trick.
    I'm not sure what we're really accomplishing by that and I believe
    it makes the rendering slower.  If we want to we could add an alt
    property, but I think we really need to rework the entire page if
    we want it to be usable without images.
  - Use &para; instead of directly encoding a paragraph char.
  - Removed classes that we no longer use like span.arrow
  - Removed the unnecessary div.header tag.
  - Added 1px of margin to the right of the <ul> element.  Before it
    was overflowing into the scrollbar area on my browser (chrome).