Fixed link text for the tasks page
I think the old logic was that we would use "Resume" if
get_subtitle_version() returned a version and if the task was assigned
to the user.

The new logic is to use "Resume" if the asignee is set.  I think this
should work fine.  get_subtitle_version() would return versions for
translation tasks before the user created a version, so we shouldn't
use that check.  The user should only see the task if it's assigned to
them or not assigned at all, so the assignee check should be the same
as before.