Fix logic for moving past the sync state (#1255)
Fixed "all lines are synced" text from always appearing when we're not
in translation mode.

Changed text to read "all lines are complete"/"not all lines are
complete".  This is more precice, since it also shows when lines are
blank in translation mode.

Refactored the Workflow class and the workflow section of our
template:
  * replaced canMoveToReview()/canMoveToTitle() with canMoveToNext().
    I like having a function that returns the correct value in both
    translation and non-translation mode.
  * replaced ng-switch to ng-if.  This seems cleaner to me and
    requires less comparisions during the digest cycle.
  * Translate the heading text for each step.
  * Added closing tags for a couple divs that didn't have them
  * Cleaned up a couple bugs/redundant code.

Updated tests and added some extra cases.