Moved dialog CSS into its own file.
This way we can share the CSS between the regular site and the
subtitle editor.

Also moved a couple other SASS pieces to their own files.

Changed the HTML/CSS structure a bit.  I think aside makes the most
sense semantically for the modal.  dialog would be awesome, but it's
only supported in some browsers.  Also, footer makes sense to me for
the buttons at the bottom.