Changed edit title/description modal to use new dialog code
Changed subtitle-editor.scss to import stuff from the shared sass
code.

Changed the way we show/hide the modal dialog.  Instead of using the
:active CSS selector, we add a shown class.  I liked the old way, but
this fixes a bug that happened when you were viewing the dialog, then
reloaded the webpage.  In that case, we would show the edit
title/description dialog over the dialog asking if you wanted to
restore your backup.

Removed the aria-hidden attribute.  If we use that we should be
updating it when we want to show the dialog, and I don't think we
were.  Also, I think simply setting display: none achieve the same
results.  Kept the aria-labeledby and role attributes.

Removed CSS for modal-style and stopped using the 3rd-party modal.scss
file.

Updated the HTML/CSS for the video title/edit icon.  I think having
the two things in the same element is nicer.