Made other dialogs use the new generic code
Added color to the CSS for dialog buttons.  Before we were relying on
inheriting the correct color.

Updated CSS for buttons not in the dialog footers for the editor.
Copied the CSS from the workflow for these.

Added code to handle multiple generic dialogs being open at once.  We
maintain a stack of the older dialogs and recreate them when the newer
generic dialog is closed.

Use ng-show instead of ng-if to hide the dialog text if it's not
present.  ng-if creates repeated elements if we try to change it
dynamically (angular issue #4852)

Refactored the idle timeout code a bit.  Added a button to the debug
modal to test it.

Removed ModalController class.  It's not needed now that we have
switched all code to the new system.