Fix error with unsynced subtitles an the timeline
There was an issue with drag and drop with unsynced subtitles.  The
code to handle DnD would use values from when we first set up the DIV
for the unsynced subtitle.  If the user had seeked since then it would
seek to a totally wrong place.

Slightly related to this was the fact that if scope.currentTime
hadn't been set yet we would set the start time for the unsynced
subtitle to null.

This may be related to #1300, I'm not sure.