Fix bug redrawing timeline.
I'm not sure why we had (this.startTime >= 0) in that if statement.
It probably made sense at one point but now it's definitely not
working.  I could make the timeline disappear by seeking to the end of
the movie, then back to the start.

This also might be related to #1300, I'm not sure.