Added dialog to save subtitles on resume session errors (#1302)
Added a new dialog for resume session errors.  It shares the same
scope with the save session error dialog (i.e. it's controlled by
SaveErrorModalController)

Removed extra </div> in the save error dialog.

Added button to show the dialog in the debug window

Changed secondsUntilClosing back to 120.  I messed with it for my last
commit.