Better UI when tasks can't be approved.
Made it so the style for a disabled approve button works (before it
was being overridden by other CSS classes).  Changed it so that we
display a pointer over the approve/send-back buttons if they're not disabled.

Changed the text for the error messages when subtitles aren't
complete.  Added a right margin to the text.