Always close dialogs when buttons are clicked.
Made to so that we always close the dialog when a button is clicked.
It's only logical to do, and we were basically doing it in every
callback.  Many times it would be the only line of code.  The one time
we didn't want to do it is when we were attempting to regain the
editor lock and in this case a freeze box makes more sense.