DialogManager.openDialog refactor
Changed a couple things with generic dialogs:
  - Moved the code that defines the title/text/buttons to the
    DialogManager class
  - Changed openDialog to input the dialog name instead of the entire
    definition.
  - Removed button callback functions from the button definition and
    made them a parameter to openDialog.

This makes dialogs more easily testable by spying on
DialogManager.openDialog.  It's much easier to check the dialog name
than the entire dialog -- and this check won't need to be updated if
we change the dialog text/title.  It's also easy to invoke callback
functions for buttons.

I'm going to be refactoring the code to save things so I wanted a way
to test that it pops up the correct dialogs.

Also I think it's a bit nicer to have the dialog definitions separate
from the code that calls them.  It's basically just a bunch of text
that overflows 80 columns, so I like having it sectioned off.  The one
thing that seems a little weird now is the overides parameter, but I
think that's an okay price to pay.  I also think I can refactor the
code that shows the subtitles saved dialog to not use overides.