One more refactor to DialogManager
Changed the way we create the DialogManager.  Instead registering a
class with the injector, we register a controller that creates a
singleton object.

I think this is the best way to handle these app-level controller
singletons.  Using a controller means that we can simply specify
dependencies like VideoPlayer and just get them passed to use rather
than force AppController to pass them along.  Also, I think this most
closely fits with the angular way, since DialogManager is a controller
class.

Made a few other changes:
  - Moved the dialog and button definitions out of DialogManager and
    into their own objects.  I think the split is nice because it
    keeps the DialogManager class lean.  It also makes it easier to
    mock out the dialogs/buttons for the tests.
  - Fixed the text for the resume button on the save dialog.
  - Added back allowClose attributes that I mistakenly deleted in the
    last commit.