Refactored saving things
Moved code that saves things into an app-level controller.  Added
tests for it.  I think this is a much better way to handle saving
things.  It also fixes the bug when exiting with unsaved changes.
Also it's much more easily tested.

Updated the HTML/CSS for the buttons on the top-right a bit.