Fix for the multi-video create subtitles form
This form is put on the tasks page and the dashboard page.  If a video
already has it's primary audio language set, we don't want to let the
user change it.  However, readonly is the correct attribute to use,
since if we use disabled then the value doesn't get POSTed at all.