Cleaned up the VideoType class
Removed the video_url() method, which was confusing and redundant.
The confusing part is that it either accept a VideoUrl object, or a
VideoType object.  The reduntant part was that VideoUrl.url and
VideoUrl.convert_to_video_url() both did the same thing.

Removed the create_kwars() method, which was mispelled and redundant.
It was only used to query VideoUrl objects, but VideoUrl can figure
out what values to pass to the query without create_kwars().

Fixed typo in the brightcove module filename.  Removed trailing
whitespace from that file.  Removed the video_data() method which was
never called.