BrightcoveVideoType update
Made it recognize brightcove URLs with the video id in the path rather
than the query string.

Changed the redirect code to use a HEAD request.  I'm not sure when
the previous note was written, but brightcove now supports HEAD
requests just fine.

Removed the convert_video_url() override function.  We no longer need
it now that the base class handles query params.