Changed way we handle multiple forms on the external accounts view
Instead of having a different HTML form for each account form, we are
going to have 1 big form that contains the settings for each form.
Made a class that handles some of the complexity of this.  Instead of
a remove button, there will be a checkbox input that enables/disables
the account.

Refactored the AccountForm class a bit to make it work with
AccountFormset.

Still need to update the view/templates to work.