Better method of detecting new items in a feed (#1445)
Before, we would remember the top link in our RSS feed, and when we
parsed it again we would stop when we saw it.  This fails pretty bad
for brightcove feeds, since they are ordered oldest-first.

Instead, just try to import all videos, but skip ones that have
already been created.

This also alters the youtube next link code handling a bit.  Now,
instead of only checking the next link on the first pass (when we
don't have last_link set), we check the next link as long as all of
the items in the feed are new.