Better way to skip importing youtube videos.
The new way is that we follow the next link on the first time we
import videos, then not again.  This matches how things used to work,
since we would skip following the next link when last_link was not
None.  I think the new way is easier to understand though, since it's
more explicit.