Added code to create youtube accounts
Copied the old youtube account code from the accountlinker app and
updated it based on youtube API changes.  In particular, now youtube
uses a channel ID to identify video owners instead of the youtube
username.

Created migration for the YouTubeAccount model.  Changed the fields a
bit to use channel IDs

Added utils.youtube module.  It's going to handle the YouTube API.
It's a standalone module since we want to use it in both the
externalsites app and the videos app (for the YouTube video type).

Changed the externalsites account form handling to work with youtube.
It's a bit different then the rest of the code since youtube uses
oauth instead of the user giving us their credentials.

Added some OAuth-related views to externalsites.

Added YOUTUBE_CLIENT_FORCE_HTTPS to settings.py and set the default to
True.  This seems much better than the old way of using getattr on the
settings module.