Updated code to get youtube video info
Changed it to use the new utils.youtube module.  Added
get_video_info() function to that module.  Made owner_username be the
channel id instead of the youtube username.  Added a migration to set
exsting usernames to NULL.

Refactored the code in the videos module that updates the
Video/VideoUrl objects based on the video type.  I think the new way
is cleaner and more consistent than the older way.  Before
set_values() would sometimes save the Video and sometimes not.  It
would also return the Video object back for basically no reason.

Removed the code to add the amara description to youtube videos.  I'm
going to replace that code with newer stuff and it makes sense to me
to do this when we're syncing.