Don't store access tokens in the DB
We are just going to have to refresh them before we do any operations
so it doesn't make sense to use the storage.

Added function to get new access tokens using the refresh token.
Refactored the utils.youtube module a touch.

Updated docker_dev_settings so that the default Site object is
hopefully correct now.