Couple of performance improvements for the sync history tab
I noticed that we were doing several queries per history row, which is
never good.  Changed it so that we select the version when we select
the sync history objects and that we select all the accounts using 1
query per account type.  Lastly, removed the link to the version, it
required several queries and it didn't seem all that useful.