Refactored search app and rewrote search tests
Cleaned up the API for SearchForm and SearchApiClass.  Fixed a bunch
of weirdness there:

  - The API of SearchForm was optionally inputting a haystack
    SearchQuerySet, but we always wanted to use that
  - For most searches we would run apply_query() twice, which resulted
    in weird search queries
  - In general there was just a bunch of redundant code

Changed the tests so that they don't use fixtures and they don't hit
the solr database.  Using solr in the tests is slow and it's easy to
mess things up if you don't clean up the state after the test.

Instead we simply test if we are creating the correct queries and
assume that haystack will do its job correctly.  Also moved basically
all the logic into the SearchForm and tested that instead of testing
the entire rpc function which does a bunch of other stuff.

Tests now run in under a second on my machine instead of about 45
seconds before.