dockerdev improvements
I'm now on version 0.9.0 and realize that it uses long cidfiles, so I
fixed the dockerdev scripts to handle that.

Use "--cidfile" instead of "-cidfile" on newer versions.

Made the version module handle determining what behavior we need to
change instead of just calculating the version number.  This makes the
other modules more readable and also makes it a bit easier to change
things.