Removed imports from the widget.tests module
We don't have that module around anymore and it was causing other tests
to fail.

Started refactoring the teams tests.  I needed to work on that
unittest anyways since it used the RequestMockup class.  I would love
to use the same technique for a bunch of the other teams tests.