Fixed fetching existing subs from youtube (#1590)
The original issue was that when we were scheduling the task to fetch
subs, the video object wasn't yet saved but we needed to have the
primary key.

I fixed that and also moved the code around.  Now the youtube API code
to fetch subtitles is in utils.youtube and the code that handles
adding the SubtitleVersions is in externalsites.

This was the last remaining bit of youtube API code in the youtube
video type module.  That module is a lot simpler now.