Fig/docker changes
Instead of using the build command in fig.yml use image.  This allows
use to use a single docker image for both app and worker.  The
downside is that there's an extra step in the install process.

Add DJANGO_SETTINGS_MODULE to the environment for our fig images.
This removes the need for an extra command line argument.