One more change for patch_reverse()
Putting patch_reverse() in settings.py is not good because
patch_reverse() needs to import the django settings.  Since
settings.py is still being imported this results in a weird situation.
In particular, we get a crash if DJANGO_SETTINGS_MODULE is not set.

Putting it in models.py seems okay since it gets loaded fairly early
on.  I would rather we just avoid the monkey patching altogether, but
this seems okay for now.