Started using staticmedia instead of unisubs_compressor in our templates
Added template tag to include a media bundle.  It doesn't support S3
yet.

Changed the URL conf for our media urls.  Now bundles are put into
the css/js subdirectories, which mimics the old structure.  Also we
now serve up the images and fonts directory in the urlconf.  This is
needed for our CSS relative paths to work (for example
url(../images/amara-logo.png)).

Removed a bunch of things that we no longer use/need:
  - pagedemo app
  - The unisubs-stat js file (It wasn't used by any of our apps, but
    it got built and uploaded to
    http://s3.www.universalsubtitles.org/js/unisubs-statwidget.js)
  - The test_form page (/videos/test)
  - The jstesting app (bunch of pages that start with /jstesting/)
  - The streamer app (bunch of pages that start with /streamer/)

Removed some javascript calls to the closure code:
  - unisubs.widget.WidgetController.makeGeneralSettings.  I think this
    is no longer needed with the new embedder.
  - unisubs.messaging.simplemessage.displayPendingMessages.  I don't
    think we use the simplemessage framework anymore.

No longer include unisubs-onsite-compiled.js I think this was only
needed for the old widget and we don't use that on our site anymore.
We were also sometimes using offsite-compiled for our own pages which
seems strange.

Added a semicolon at the end of jquery.chosen.min.js.  I guess it left
it out to save the byte, but this doesn't work when we combine
javascript files together.

Still to do:
  - new editor JS
  - embedder (both JS and CSS)
  - old embedder
  - widgitizer