Changed the editor to use the new media system.
Used ng-annotate to add dependency annotations for our angular code.
This is needed because the subtitle editor JS is now minified.

Also use the new media system for the editor CSS.  This means no more
compiling subtitle-editor-dev.css by hand and checking it in!  Updated
the URLs for our editor CSS so they are now relative to the media/css
directory.