Added management command to build javascript bundles with closure
The closure build process is pretty crazy and we are in the process of
moving away from using it.  This command is a semi-hack to just allow
us to keep up support for them.  The idea is that hopefully we will
just let these files site there and continue to deploy them, but not
really change anywith.  In case we do want to change something, then
we need rebuild that bundle and check the result in to git.

This is all in preparation of adding support for the old embedder, old
editor, and widgetizer to the staticmedia app.

Moved a bunch of javascript settings out of our main settings file and
into the build_closure_bundle module.  Moved the javascript paths for
editor.js to live inside the MEDIA_BUNDLE setting.  This seems more
consistant with the way the other bundles are defined.