Made the embedder work with the staticmedia app
Added support for adding the _amaraConf object to javascript files.
This replaces our old system where we would either include
embedder/conf.js or embedder/conf-dev.js when building the embedder
bundle.

Don't put the embedder JS/CSS in the release/public directory.  We
don't actually publish those URLs, instead we publish the link to the
iframe URL.  This lets us put them in the same location as other
JS/CSS bundles which is simpler.  It also allows us to drop the
bootloader code that was happening in release/public/embedder.js (but
only on production).

Updated the URLs in embedder.scss based on the new location.  Changed
embedder.js so that the init() call happens inside
$(document).ready.  This is needed now because that file is loaded
directly instead of with a bootloader.

Don't have embedder.js add the embedder.css script dynamically.  Since
this is all happening in an iframe that we control, we can simply
include the CSS in the HTML head.

Removed embedder-dev.css, we now just build it on the fly.

Removed embedder-offsite-iframe.html.  I don't think we use it
anymore.