Made the embedder work with HTTPS
Added a page to test the embedder (/media/test/embedder/)

Added a static_url() template tag.

Fixed our embedder JS so that it uses the same protocol as the request
for URLs instead of forcing HTTP.  For most URLs this simply meant
changing "http://" to "//" (AKA a protocol-relative URL).  For some
reason this didn't work for iframeDomain, but the fix was simple
enough there, just use parser.protocol.