Don't redirect users for the youtube callback
Before this change we would have an extra redirect when adding a
youtube account because of localeurl.  The browser would navigate
these pages:

  1) /en/profiles/account/
  2) https://accounts.google.com/o/oauth2/auth
  3) /externalsites/youtube-callback
  4) /en/externalsites/youtube-callback
  5) /en/profiles/account/

Since the youtube callback doesn't need to be localized, we can skip
step 4.  The change to LOCALE_INDEPENDENT_PATHS accomplishes this.
Also we can remove the old entry there for the accountlinker path.