Added nicer way to run dev commands
Made a new directory that contains commands to do common tasks with
the dev environment.

The system works similarly to virtualenv.  You source the activate
script, which adds the directory to your path.  When you're done you
run deactivate.

Features include:
  - You can run the commands from any directory
  - Less typing
  - Quick way to start up a bash prompt

Also removed a couple scripts that had to do with vagrant.  I don't
think we plan on supporting that stuff anymore.