Add analytics to the subtitle editor
Moved the analytics code block into a separate template and included
that from subtitle-editor.html.

Simplified the cases for when we don't use analytics because we're
running from dev or something.  Before we sometimes would set up the
code, but not include the script.  Or we might include the github CSS
and highlight JS.  Both of these things seemed pretty weird, now
we either show the analytics, or show a comment instead.