Fixes for a bunch of included media files
With the new media changes, we can only include media bundles, not
just link to anything in the media/js and media/css directory.  Fixed
a bunch of links using the old method.  Added media bundles for:

  - ajax paginator (used on the videos/search pages).
  - prepopulate.js (used on the team create page).
  - ie8.css (for IE 8 and lower).

Removed some things that used the old linkage method that we don't use
anymore:
  - volunteer page.
  - JS/CSS in the statistic page.
  - Widget JS in the FAQ and About pages.
  - jquery.metadata.js in the messages pages (We include that file in
    site.js anyway).

The one remaining place is the widget directory, but I don't really
understand what's going on there.  I want to just leave that alone
until we ditch the old embedder and editor and then say goodbye to the
widget altogether.

It would be nice to have the original non-minified sources for some of
these JS files like prepopulate.js, but I have no clue where they came
from.