Add the credit test even if subtitles are not complete
I believe this matches the old behavior, but in any case it makes
sense.  The text is a request to subtitle the video on amara.
So even if there are not subtitles yet, having the link still makes
sense.  Now that I think about it "credit" is really the best word for
this.