Remove assertion check (#1649)
I can't reproduce this on my local instance and I think this check is
not really good.  There are a lot of ways for python equality to be
false which wouldn't mean that there are data integrity issues.
Throwing an exception there could create issues actually, so let's
just not do the check.

If this was something that we should be worried about, then a unit
test seems much more appropriate.