Don't try to import accountlinker accounts if channel_id is not set
When we have run the command to lookup the channel ids, it fails for
many of the rows in our database.  I believe this is because the user
has either revoked access to amara, deleted/altered their account, or
done something similar.  We should just forget about the account and
ask the user to re-link it if they want to.