Add allow_sync_teams field to YouTubeAccount
This is based off a pending change where we will only sync to team
YouTube accounts for video's owned by that team.  One issue with that
system is when one group of users controls multiple teams (for example
a prep team and a delivery team).  This field will allow the admins to
add those additional teams to sync.

Updated the docs to reflect the new system (which hasn't been totally
implemented yet).

Updated the migrations a bit:
  - Added the allowed_sync_teams field to
    0006_create_youtube_account_table
  - Updated a couple migrations to fix the unique_together constraint
    on type/owner id for Kaltura.  This is just a change to the
    database schema that is stored for South.  We were always actually
    creating that constraint.  For some reason the south schema was
    correct for Brightcove, but not Kaltura.

Changed models.py so that the unique_together constraints are in the
BrightcoveAccount/KalturaAccount classes.  Before they were in the
base class, but that wasn't working correctly.  Not sure if it was a
bug in our code or south, but the direct way fixes the issue.